refactor(summariser): rename summarizer to summariser for consistency
Standardise the naming of the summariser module to use British English spelling (summariser) instead of American English (summarizer) for consistency across the codebase. This change affects both the package name and related function calls.
